requirements.txt
streamlit
langchain
langchain-openai
langchain-community
openai
msal
requests
python-dotenv
chromadb
tiktoken